requirements.txt
absl-py==0.9.0
astor==0.8.1
cycler==0.10.0
decorator==4.4.1
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
imageio==2.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.16.4
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
